Add correct finished process.

+++ start.sh
@@ -11,31 +11,34 @@ curl -s https://www.ethercalc.org/${FINISHED}.csv | tr -d ',' > queue-finished.t
 # get lastest landsat filename to process
 NAME=`head -n1 queue-pending.txt`
 
-# download landsat
+# 1. download landsat
 if [ ! -f ~/landsat/zip/${NAME}.tar.bz ]; then
   echo "Downloading ${NAME} ..."
   landsat download ${NAME}
 fi
 
-# download completed
-if [ -f ~/landsat/zip/${NAME}.tar.bz ]; then
+# 2. download completed, processing
+if [ -f ~/landsat/zip/${NAME}.tar.bz ] && [ ! -f ~/landsat/processed/${NAME}/final-pan.TIF ] ; then
   # update queue list
-  echo "$(tail -n +2 queue-pending.txt)" > queue-pending.txt
   echo "${NAME}" >> queue-downloaded.txt
-  sed 's/$/,,/g' queue-pending.txt > queue-pending.csv
   sed 's/$/,,/g' queue-downloaded.txt > queue-downloaded.csv
-  curl -X PUT -H 'Content-Type: text/csv' --data-binary @queue-pending.csv https://www.ethercalc.org/_/${PENDING}
   curl -X PUT -H 'Content-Type: text/csv' --data-binary @queue-downloaded.csv https://www.ethercalc.org/_/${DOWNLOADED}
 
   # image process, pansharping
   echo "Image processing ${NAME} ..."
   landsat process --pansharpen ~/landsat/zip/${NAME}.tar.bz
 
+  # TODO: NDVI version of pan
   # TODO: tilelize file
-  # TODO: upload image which successful processing
+fi
 
-  # after long time process
+# 3. finished processing
+if [ -f ~/landsat/processed/${NAME}/final-pan.TIF ]; then
+  # TODO: upload image which successful processing
   echo "Writing finish record for ${NAME} ..."
+  echo "$(tail -n +2 queue-pending.txt)" > queue-pending.txt
+  sed 's/$/,,/g' queue-pending.txt > queue-pending.csv
+  curl -X PUT -H 'Content-Type: text/csv' --data-binary @queue-pending.csv https://www.ethercalc.org/_/${PENDING}
   echo "${NAME}" >> queue-finished.txt
   sed 's/$/,,/g' queue-finished.txt > queue-finished.csv
   curl -X PUT -H 'Content-Type: text/csv' --data-binary @queue-finished.csv https://www.ethercalc.org/_/${FINISHED}
